+++ packages/common/util.js
@@ -19,6 +19,8 @@ const path = require('path');
 const mime = require('mime-types');
 const log = require('./log');
 
+const utils = require('@cumulus/aws-client/utils');
+
 /**
  * Mark a piece of code as deprecated
  *
@@ -141,11 +143,8 @@ exports.isNil = (x) => exports.isNull(x) || exports.isUndefined(x);
  * @param {string} newStack - a stack trace
  */
 exports.setErrorStack = (error, newStack) => {
-  // eslint-disable-next-line no-param-reassign
-  error.stack = [
-    error.stack.split('\n')[0],
-    ...newStack.split('\n').slice(1)
-  ].join('\n');
+  exports.deprecate('@cumulus/common/util/setErrorStack', '1.17.1', '@cumulus/aws-client/utils/setErrorStack');
+  utils.setErrorStack(error, newStack);
 };
 
 /**